pyproject.toml
[tool.poetry]
name = "seeker-backend"
version = "0.1.0"
description = ""
authors = ["Clooooode <clode@clo5de.info>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.8"
kink = "^0.8.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
google-api-python-client = "^2.160.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
pymongo = "^4.11"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"
pre-commit = "^4.1.0"
isort = "^6.0.0"

[tool.ruff]
target-version = "py313"
exclude = ["src/script/"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  # "UP",  # pyupgrade
]
ignore = [
  "E501", # line too long
  "B008",  # do not perform function calls in argument defaults
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
